% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a1_heading.R
\name{a1_heading}
\alias{a1_heading}
\title{Format A1 cell in every sheet to Heading style for accessibility purposes}
\usage{
a1_heading(file)
}
\arguments{
\item{file}{name of an ods file to add heading format to}
}
\description{
Format A1 cell in every sheet to Heading style for accessibility purposes
}
