% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_renv.R
\name{check_package_v}
\alias{check_package_v}
\title{Check and Update Package Version}
\usage{
check_package_v(package_deets, problem_libs, lockfile_path)
}
\arguments{
\item{package_deets}{The details of the package from the `renv` lockfile.}

\item{problem_libs}{The data frame containing the list of problem packages and their target versions.}

\item{lockfile_path}{The path to the `renv` lockfile.}
}
\value{
Messages indicating whether the package was updated or already at the correct version.
}
\description{
This function checks if the version of a package in the `renv` lockfile is less than the specified target version.
If it is, the package version is updated in the lockfile.
}
