% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Remove creator names from the metadata of an ods file and add new tags if wanted}
\usage{
clean_metadata(file, tags = NULL)
}
\arguments{
\item{file}{name of an ods file to edit the metadata}

\item{tags}{optional tags to add to the file. Defaults to NULL}
}
\description{
Remove creator names from the metadata of an ods file and add new tags if wanted
}
