% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ods.R
\name{convert_to_ods}
\alias{convert_to_ods}
\title{Convert an Excel file to the equivalent ods object in Cloud R.
Note that this function will not work in local versions of R.}
\usage{
convert_to_ods(file, folder = getwd())
}
\arguments{
\item{file}{name of an xlsx file to convert to ods}

\item{folder}{directory location of the xlsx file, and where the ods file
will save to. Defaults to the current working directory}
}
\description{
Convert an Excel file to the equivalent ods object in Cloud R.
Note that this function will not work in local versions of R.
}
