% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sharepoint.R
\name{export_sharepoint}
\alias{export_sharepoint}
\title{Export an R object as a CSV or XLSX file to a specified Sharepoint location}
\usage{
export_sharepoint(x, file, site, drive)
}
\arguments{
\item{x}{data object in R}

\item{file}{Name and location of the file you want to save to sharepoint. Includes the extension and any folders after the drive name}

\item{site}{Name of sharepoint site you want to save a file to. Can be found in the URL of the Sharepoint location, e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail", the site name would be Rail}

\item{drive}{Name of the sharepoint drive you want to save the file to Can be found in the URL of the sharepoint location e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail/RailStats", the drive name is "RailStats". Can also be found using the list_sharepoint_drives() function in this package.}
}
\description{
Export an R object as a CSV or XLSX file to a specified Sharepoint location
}
