% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sharepoint.R
\name{import_sharepoint}
\alias{import_sharepoint}
\title{Import a file from a specified Sharepoint location}
\usage{
import_sharepoint(file, site, drive, destination = NULL)
}
\arguments{
\item{file}{Name and location of the file you want to extract from sharepoint. Includes the extension and any folders after the drive name}

\item{site}{Name of sharepoint site you want to take a file from. Can be found in the URL of the Sharepoint location, e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail", the site name would be Rail}

\item{drive}{Name of the sharepoint drive the file sits on. Can be found in the URL of the sharepoint location e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail/RailStats", the drive name is "RailStats". Can also be found using the list_sharepoint_drives() function in this package.}

\item{destination}{Folder and filename location to download the file to. Defaults to NULL, which saves the file with it's current filename in your current working directory.}
}
\description{
Import a file from a specified Sharepoint location
}
