% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_left.R
\name{left}
\alias{left}
\title{Equivalent of left in SQL; takes the specified number of characters from
the left hand side of a string.}
\usage{
left(x, n = 1)
}
\arguments{
\item{x}{a string or vector of strings}

\item{n}{integer. Number of characters you want to extract from the string.}
}
\description{
Equivalent of left in SQL; takes the specified number of characters from
the left hand side of a string.
}
