% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sharepoint.R
\name{list_sharepoint_drives}
\alias{list_sharepoint_drives}
\title{List the sharepoint drives available in the specified site, useful for finding their names}
\usage{
list_sharepoint_drives(site)
}
\arguments{
\item{site}{Name of sharepoint site you want to take a file from. Can be found in the URL of the Sharepoint location, e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail", the site name would be Rail}
}
\description{
List the sharepoint drives available in the specified site, useful for finding their names
}
