% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_file.R
\name{recent_file}
\alias{recent_file}
\title{Find the most recently modified file in a folder
Can use a matching pattern to specify a specific file type or name.}
\usage{
recent_file(filepath, pattern = NULL, recursive = FALSE, ignore.case = FALSE)
}
\arguments{
\item{filepath}{folder to search in to find most recent file}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}

\item{recursive}{logical. Should the listing look inside subfolders in
the filepath? Defaults to FALSE.}

\item{ignore.case}{logical. Should pattern-matching be case-insensitive?
Defaults to FALSE}
}
\description{
Find the most recently modified file in a folder
Can use a matching pattern to specify a specific file type or name.
}
