% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_colours.R
\name{remove_colours}
\alias{remove_colours}
\title{Remove font colours from ODS files to ensure they only use automatic colours (exception for hyperlinks)}
\usage{
remove_colours(file)
}
\arguments{
\item{file}{name of an ods file to remove font colours from}
}
\description{
Remove font colours from ODS files to ensure they only use automatic colours (exception for hyperlinks)
}
