% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_notes.R
\name{remove_notes}
\alias{remove_notes}
\title{Remove notes from DfT publication table headers}
\usage{
remove_notes(data, trim = TRUE)
}
\arguments{
\item{data}{Data frame with column names}

\item{trim}{Logical - TRUE to trim whitespace after removing notes, otherwise FALSE, default is TRUE.}
}
\description{
This will have no effect if the data frame does not contain notes in the header.
The function identifies a note as anything inside square brackets [], check the number of notes removed in the message is as you expect.
}
