% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_left.R
\name{right}
\alias{right}
\title{Equivalent of right in SQL; takes the specified number of characters from
the right hand side of a string.}
\usage{
right(x, n = 1)
}
\arguments{
\item{x}{a string or vector of strings}

\item{n}{integer. Number of characters you want to extract from the string.
Defaults to 1.}
}
\description{
Equivalent of right in SQL; takes the specified number of characters from
the right hand side of a string.
}
