% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlConnect.R
\name{sqlConnect}
\alias{sqlConnect}
\title{Connect to a SQL database on the Cloud R computing environment}
\usage{
sqlConnect(server, database)
}
\arguments{
\item{server}{string. Name of the server you would like to connect to.}

\item{database}{string. Name of the database within the server you would like
to connect to}
}
\description{
Connect to a SQL database on the Cloud R computing environment
}
