% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_renv.R
\name{update_renv}
\alias{update_renv}
\title{Update Renv Lockfile}
\usage{
update_renv(lockfile_path = "renv.lock")
}
\arguments{
\item{lockfile_path}{Character. The path to the `renv` lockfile. Defaults to `"renv.lock"`, the lockfile in your current project..}
}
\value{
A message indicating which problem packages were identified, and their updated versions
}
\description{
This function checks the versions of packages in an `renv` lockfile against a list of known packages which are incompatible with the upgraded version of R.
If any problem packages are found, they are replaced with appropriate versions in your lock file.
}
