% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clipboard.R
\name{writeClipboard}
\alias{writeClipboard}
\title{Copy data to an appropriately-sized clipboard where available}
\usage{
writeClipboard(df)
}
\arguments{
\item{df}{a dataframe object}
}
\description{
Copy data to an appropriately-sized clipboard where available
}
